app/consumer/requirements.txt
confluent_kafka
pyyaml
requests